Move remaining crypto classes into superpackage
