+++ core/src/main/kotlin/net/corda/core/contracts/Amount.kt
@@ -1,6 +1,6 @@
 package net.corda.core.contracts
 
-import net.corda.core.crypto.composite.CompositeKey
+import net.corda.core.crypto.CompositeKey
 import net.corda.core.identity.Party
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.utilities.exactAdd
