+++ core/src/main/kotlin/net/corda/core/crypto/CompositeKeyFactory.kt
@@ -1,4 +1,4 @@
-package net.corda.core.crypto.composite
+package net.corda.core.crypto
 
 import java.security.*
 import java.security.spec.InvalidKeySpecException
