+++ core/src/main/kotlin/net/corda/core/crypto/CompositeSignature.kt
@@ -1,7 +1,6 @@
-package net.corda.core.crypto.composite
+package net.corda.core.crypto
 
 import net.corda.core.crypto.SecureHash
-import net.corda.core.crypto.TransactionSignature
 import net.corda.core.serialization.deserialize
 import java.io.ByteArrayOutputStream
 import java.security.*
