+++ core/src/main/kotlin/net/corda/core/crypto/CryptoUtils.kt
@@ -2,7 +2,6 @@
 
 package net.corda.core.crypto
 
-import net.corda.core.crypto.composite.CompositeKey
 import net.corda.core.utilities.OpaqueBytes
 import java.math.BigInteger
 import java.security.*
