+++ core/src/test/kotlin/net/corda/core/contracts/TransactionTests.kt
@@ -1,7 +1,7 @@
 package net.corda.core.contracts
 
 import net.corda.core.crypto.*
-import net.corda.core.crypto.composite.CompositeKey
+import net.corda.core.crypto.CompositeKey
 import net.corda.core.identity.Party
 import net.corda.core.transactions.LedgerTransaction
 import net.corda.core.transactions.SignedTransaction
