+++ node/src/main/kotlin/net/corda/node/utilities/ServiceIdentityGenerator.kt
@@ -1,6 +1,6 @@
 package net.corda.node.utilities
 
-import net.corda.core.crypto.composite.CompositeKey
+import net.corda.core.crypto.CompositeKey
 import net.corda.core.crypto.generateKeyPair
 import net.corda.core.identity.Party
 import net.corda.core.utilities.loggerFor
